Split train HUD input and route preview responsibilities
