+++ moorestech_client/Assets/Scripts/Client.Game/InGame/UI/UIState/State/TrainHUDScreen/TrainRidingInputSender.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 2ea56fc4e07816a4990ddddea0d66502
