+++ moorestech_client/Assets/Scripts/Client.Game/InGame/UI/UIState/State/TrainHUDScreenState.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Collections.Generic;
 using System.Threading;
 using Client.Game.InGame.Context;
 using Client.Game.InGame.Control;
@@ -10,9 +9,6 @@ using Client.Game.InGame.UI.UIState.State.PauseMenu;
 using Client.Game.InGame.UI.UIState.State.TrainHUDScreen;
 using Cysharp.Threading.Tasks;
 using Game.PlayerRiding.Interface;
-using Game.Train.RailCalc;
-using Game.Train.RailGraph;
-using Game.Train.Unit;
 using MessagePack;
 using Server.Event.EventReceive;
 using Server.Protocol.PacketResponse;
@@ -20,33 +16,18 @@ using UnityEngine;
 
 namespace Client.Game.InGame.UI.UIState.State
 {
-    // 列車に乗車中の HUD ステート。State が列車関連処理の唯一の起点である
-    // Train HUD state, the single source of truth for all train-related processing while riding.
+    // 列車に乗車中の HUD ステート。状態遷移と列車系UI処理の呼び出しを担当する
+    // Train HUD state, responsible for state transitions and dispatching train UI work while riding.
     public class TrainHUDScreenState : IUIState
     {
-        private const float TrainInputHeartbeatIntervalSeconds = 2f;
-        private const int BranchRoutePreviewSearchNodeLimit = 3;
-        private const int BranchRoutePreviewSamplesPerSegment = 18;
-        private const float BranchRoutePreviewHeightOffset = 0.22f;
-        private const float BranchRoutePreviewWidth = 0.18f;
-        private static readonly Color BranchRoutePreviewColor = new(1f, 0.78f, 0.18f, 0.92f);
-
         private readonly PlayerStateController _playerStateController;
         private readonly TrainUnitClientCache _trainUnitClientCache;
         private readonly TrainHudScreenUIStateController _subStateController;
-        private readonly List<IRailNode> _branchRouteNodes = new();
-        private readonly List<IRailNode> _branchCandidateNodes = new();
-        private readonly List<Vector3> _branchRoutePoints = new();
+        private readonly TrainRidingInputSender _trainRidingInputSender = new();
+        private readonly TrainBranchRoutePreviewController _branchRoutePreviewController = new();
 
         private bool _isDismountTrain = false;
         private RidingPlayerStateContext _rideContext;
-        private bool _hasSentTrainMoveInput;
-        private bool _lastSentMoveForward;
-        private bool _lastSentMoveBackward;
-        private float _lastTrainInputSentAt;
-        private GameObject _branchRoutePreviewObject;
-        private LineRenderer _branchRoutePreviewLine;
-        private Material _branchRoutePreviewMaterial;
 
         private IDisposable _eventSubscription;
         private CancellationTokenSource _cts;
@@ -71,10 +52,7 @@ namespace Client.Game.InGame.UI.UIState.State
             
             _rideContext = null;
             _isDismountTrain = false;
-            _hasSentTrainMoveInput = false;
-            _lastSentMoveForward = false;
-            _lastSentMoveBackward = false;
-            _lastTrainInputSentAt = 0f;
+            _trainRidingInputSender.Reset();
             
             // 初期値として乗車完了済みの場合は即時反映
             // If the player is already riding at the time of entering, reflect that immediately.
@@ -140,7 +118,11 @@ namespace Client.Game.InGame.UI.UIState.State
 
         public UITransitContext GetNextUpdate()
         {
-            if (_isDismountTrain) return new UITransitContext(UIStateEnum.GameScreen);
+            if (_isDismountTrain)
+            {
+                _branchRoutePreviewController.Hide();
+                return new UITransitContext(UIStateEnum.GameScreen);
+            }
 
             // まだ乗車が完了していないのであれば何もしない
             // If riding is not yet completed, do nothing.
@@ -150,7 +132,7 @@ namespace Client.Game.InGame.UI.UIState.State
             // Force dismount if the target car has disappeared.
             if (!_trainUnitClientCache.TryGetCarSnapshot( _rideContext.CurrentCarId, out var ridingTrainUnit, out _, out _, out _))
             {
-                HideBranchRoutePreview();
+                _branchRoutePreviewController.Hide();
                 return new UITransitContext(UIStateEnum.GameScreen);
             }
                 
@@ -159,11 +141,11 @@ namespace Client.Game.InGame.UI.UIState.State
             _subStateController.Update();
             if (_subStateController.CurrentState != TrainHudScreenUIStateEnum.GameScreen)
             {
-                HideBranchRoutePreview();
+                _branchRoutePreviewController.Hide();
                 return null;
             }
 
-            UpdateBranchRoutePreview(ridingTrainUnit);
+            _branchRoutePreviewController.Update(ridingTrainUnit);
 
             // GameScreenだけ降車処理、列車操作入力を受け付け
             // Only process dismount and train control input on the GameScreen.
@@ -171,28 +153,8 @@ namespace Client.Game.InGame.UI.UIState.State
             {
                 SendDismountRequestAsync().Forget(LogRpcFault);
             }
-            
-            var moveForward = UnityEngine.Input.GetKey(KeyCode.W);
-            var selectPreviousBranch = UnityEngine.Input.GetKeyDown(KeyCode.A);
-            var moveBackward = UnityEngine.Input.GetKey(KeyCode.S);
-            var selectNextBranch = UnityEngine.Input.GetKeyDown(KeyCode.D);
-            var moveForwardChanged = !_hasSentTrainMoveInput || moveForward != _lastSentMoveForward;
-            var moveBackwardChanged = !_hasSentTrainMoveInput || moveBackward != _lastSentMoveBackward;
-            var isHeartbeatDue = _hasSentTrainMoveInput && Time.realtimeSinceStartup - _lastTrainInputSentAt >= TrainInputHeartbeatIntervalSeconds;
-            var isInput = moveForwardChanged || moveBackwardChanged || isHeartbeatDue || selectPreviousBranch || selectNextBranch;
-            if (isInput)
-            {
-                ClientContext.VanillaApi.SendOnly.SendTrainCarRidingInput(
-                    moveForward,
-                    moveBackward,
-                    selectPreviousBranch,
-                    selectNextBranch);
-                _hasSentTrainMoveInput = true;
-                _lastSentMoveForward = moveForward;
-                _lastSentMoveBackward = moveBackward;
-                _lastTrainInputSentAt = Time.realtimeSinceStartup;
-            }
-            
+
+            _trainRidingInputSender.Update();
 
             return null;
 
@@ -235,7 +197,7 @@ namespace Client.Game.InGame.UI.UIState.State
             // ステートを変更して降車処理を実行
             // Change state to trigger dismount processing.
             _playerStateController.SetState(PlayerStateEnum.Normal, null);
-            DestroyBranchRoutePreview();
+            _branchRoutePreviewController.Destroy();
 
         }
 
@@ -244,148 +206,6 @@ namespace Client.Game.InGame.UI.UIState.State
         private static void LogRpcFault(Exception exception)
         {
             Debug.LogWarning($"[TrainHUDScreenState] RPC fault: {exception}");
-        }
-
-        private void UpdateBranchRoutePreview(ClientTrainUnit trainUnit)
-        {
-            if (!TryBuildBranchRoutePreviewPoints(trainUnit))
-            {
-                HideBranchRoutePreview();
-                return;
-            }
-
-            EnsureBranchRoutePreviewRenderer();
-            _branchRoutePreviewObject.SetActive(true);
-            _branchRoutePreviewLine.positionCount = _branchRoutePoints.Count;
-            for (var i = 0; i < _branchRoutePoints.Count; i++)
-            {
-                _branchRoutePreviewLine.SetPosition(i, _branchRoutePoints[i]);
-            }
-        }
-
-        private bool TryBuildBranchRoutePreviewPoints(ClientTrainUnit trainUnit)
-        {
-            _branchRouteNodes.Clear();
-            _branchRoutePoints.Clear();
-            var railPosition = trainUnit?.RailPosition;
-            if (railPosition == null) return false;
-
-            // 現在の進行先から3node以内で最初の分岐を探す。
-            // Search the first branch within three nodes from the current route.
-            var previousNode = railPosition.GetNodeJustPassed();
-            var currentNode = railPosition.GetNodeApproaching();
-            if (currentNode == null) return false;
-            var firstSegmentStart = ResolveFirstSegmentStart(railPosition, previousNode, currentNode);
-            if (previousNode != null) _branchRouteNodes.Add(previousNode);
-            _branchRouteNodes.Add(currentNode);
-
-            if (!TryAppendRouteToNextBranch(previousNode, currentNode, trainUnit.GetManualBranchSelectionIndex())) return false;
-
-            // 現在位置から分岐先まで、レール曲線に沿ったLineRenderer点列を作る。
-            // Build LineRenderer points along the rail curve from current position to the selected branch.
-            BuildBranchRoutePoints(firstSegmentStart);
-            return _branchRoutePoints.Count >= 2;
-        }
-
-        private float ResolveFirstSegmentStart(global::Game.Train.RailPositions.RailPosition railPosition, IRailNode previousNode, IRailNode currentNode)
-        {
-            if (previousNode == null) return 0f;
-            var segmentDistance = previousNode.GetDistanceToNode(currentNode);
-            if (segmentDistance <= 0) return 0f;
-            return Mathf.Clamp01(1f - railPosition.GetDistanceToNextNode() / (float)segmentDistance);
-        }
-
-        private bool TryAppendRouteToNextBranch(IRailNode previousNode, IRailNode currentNode, int branchSelectionIndex)
-        {
-            for (var nodeIndex = 0; nodeIndex < BranchRoutePreviewSearchNodeLimit; nodeIndex++)
-            {
-                CopyConnectedNodes(currentNode);
-                if (_branchCandidateNodes.Count >= 2)
-                {
-                    var selectedNode = TrainUnitBranchSelector.SelectManualBranchNode(previousNode, currentNode, _branchCandidateNodes, branchSelectionIndex);
-                    _branchRouteNodes.Add(selectedNode);
-                    return true;
-                }
-
-                // 分岐ではない単一路線だけを先へ進む。
-                // Continue only through a single non-branch route.
-                if (_branchCandidateNodes.Count != 1) return false;
-                var nextNode = _branchCandidateNodes[0];
-                if (nextNode == null || nextNode == previousNode) return false;
-                previousNode = currentNode;
-                currentNode = nextNode;
-                _branchRouteNodes.Add(currentNode);
-            }
-            return false;
-        }
-
-        private void CopyConnectedNodes(IRailNode node)
-        {
-            _branchCandidateNodes.Clear();
-            foreach (var connectedNode in node.ConnectedNodes)
-            {
-                _branchCandidateNodes.Add(connectedNode);
-            }
-        }
-
-        private void BuildBranchRoutePoints(float firstSegmentStart)
-        {
-            for (var i = 0; i < _branchRouteNodes.Count - 1; i++)
-            {
-                var startT = i == 0 ? firstSegmentStart : 0f;
-                AppendBezierSegmentPoints(_branchRouteNodes[i], _branchRouteNodes[i + 1], startT);
-            }
-        }
-
-        private void AppendBezierSegmentPoints(IRailNode fromNode, IRailNode toNode, float startT)
-        {
-            BezierUtility.BuildRenderControlPoints(fromNode.FrontControlPoint, toNode.BackControlPoint, out var p0, out var p1, out var p2, out var p3);
-            for (var i = 0; i <= BranchRoutePreviewSamplesPerSegment; i++)
-            {
-                if (i == 0 && _branchRoutePoints.Count > 0) continue;
-                var t = Mathf.Lerp(startT, 1f, i / (float)BranchRoutePreviewSamplesPerSegment);
-                _branchRoutePoints.Add(BezierUtility.GetBezierPoint(p0, p1, p2, p3, t) + Vector3.up * BranchRoutePreviewHeightOffset);
-            }
-        }
-
-        private void EnsureBranchRoutePreviewRenderer()
-        {
-            if (_branchRoutePreviewLine != null) return;
-
-            // 警告色の赤ではなく、選択ルートとして読める黄橙のワールドラインを使う。
-            // Use an amber world-space line so it reads as route selection rather than danger.
-            _branchRoutePreviewObject = new GameObject("Train Branch Route Preview");
-            _branchRoutePreviewLine = _branchRoutePreviewObject.AddComponent<LineRenderer>();
-            _branchRoutePreviewLine.useWorldSpace = true;
-            _branchRoutePreviewLine.startWidth = BranchRoutePreviewWidth;
-            _branchRoutePreviewLine.endWidth = BranchRoutePreviewWidth;
-            _branchRoutePreviewLine.numCornerVertices = 4;
-            _branchRoutePreviewLine.numCapVertices = 4;
-            _branchRoutePreviewLine.shadowCastingMode = UnityEngine.Rendering.ShadowCastingMode.Off;
-            _branchRoutePreviewLine.receiveShadows = false;
-
-            var shader = Shader.Find("Sprites/Default") ?? Shader.Find("Unlit/Color");
-            _branchRoutePreviewMaterial = new Material(shader);
-            _branchRoutePreviewMaterial.color = BranchRoutePreviewColor;
-            _branchRoutePreviewLine.material = _branchRoutePreviewMaterial;
-            _branchRoutePreviewLine.startColor = BranchRoutePreviewColor;
-            _branchRoutePreviewLine.endColor = BranchRoutePreviewColor;
-        }
-
-        private void HideBranchRoutePreview()
-        {
-            if (_branchRoutePreviewLine == null) return;
-            _branchRoutePreviewLine.positionCount = 0;
-            _branchRoutePreviewObject.SetActive(false);
-        }
-
-        private void DestroyBranchRoutePreview()
-        {
-            if (_branchRoutePreviewObject != null) UnityEngine.Object.Destroy(_branchRoutePreviewObject);
-            if (_branchRoutePreviewMaterial != null) UnityEngine.Object.Destroy(_branchRoutePreviewMaterial);
-            _branchRoutePreviewObject = null;
-            _branchRoutePreviewLine = null;
-            _branchRoutePreviewMaterial = null;
         }
     }
 }
